+++ Assets/SRP/CHANGELOG.md
@@ -3,3 +3,12 @@ base
 ### 1.0.1 
 - setting path changed
 - namespace cleared
+### 2.0.0
+- changed to collection strokes
+- class name changed (Pick->Logic)
+- new module - Picker
+- types changed to string to support building
+- listDrawer -> pickableList
+- arrayDrawer -> pickableArray
+- added componentRectPicker
+- serializer extracted